# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FTeam Catalog API",
        "description": "Catálogo, sincronização e estatísticas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor da aplicação"
        }
    ],
    "paths": {
        "/api/stats": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Estatísticas agregadas",
                "operationId": "86c880c56e9da2f2547024e53ac8fde6",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "integer"
                                        },
                                        "total_products": {
                                            "type": "integer"
                                        },
                                        "total_categories": {
                                            "type": "integer"
                                        },
                                        "min_price": {
                                            "type": "number",
                                            "format": "float",
                                            "nullable": true
                                        },
                                        "max_price": {
                                            "type": "number",
                                            "format": "float",
                                            "nullable": true
                                        },
                                        "avg_price": {
                                            "type": "number",
                                            "format": "float",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/integrations/fakestore/sync": {
            "post": {
                "tags": [
                    "Sync"
                ],
                "summary": "Dispara a sincronização (sincrona ou assíncrona via ?async=1)",
                "operationId": "21537c09f7f9b21fa8a7516d19600310",
                "parameters": [
                    {
                        "name": "async",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "imported": {
                                            "type": "integer"
                                        },
                                        "updated": {
                                            "type": "integer"
                                        },
                                        "skipped": {
                                            "type": "integer"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "client_id": {
                "type": "apiKey",
                "description": "Informe o identificador do cliente no header X-Client-Id",
                "name": "X-Client-Id",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Catalog",
            "description": "Catalog"
        },
        {
            "name": "Sync",
            "description": "Sync"
        }
    ]
}